style: move in-file test bodies to src/<module>/tests.rs across the workspace
21 modules: claim, curves, scale, wire (kalico-ethercat-rt); kalico_native,
unix_native_conn (kalico-host-rt); lib tests (kalico-protocol); drain,
enqueue, logging/{mod,context,layer,schema,writer}, motion_node,
pump/{sched_tests,tests,wire_sink_tests} (motion-bridge); log_codes,
piece_ring, step_queue (runtime). Declaration-only #[cfg(test)] mod stays
in-file; bodies match the compat/gcode convention. 1416/1416 tests
unchanged; clippy (incl. --features hw) and fmt clean.
