+++ rust/motion-bridge/src/pump.rs
@@ -168,393 +168,10 @@ pub fn schedule(
 }
 
 #[cfg(test)]
-mod sched_tests {
-    use super::*;
-
-    /// Build a queue from (tick_start, host_time) pairs.
-    fn q_with_host(ring_depth: u32, starts: &[(u64, f64)]) -> AxisQueue {
-        let mut q = AxisQueue::new(ring_depth);
-        for &(s, h) in starts {
-            q.pieces.push_back((
-                PieceEntry {
-                    start_time: s,
-                    coeffs: [0.0; 4],
-                    duration: 0.001,
-                    _reserved: 0,
-                },
-                h,
-            ));
-        }
-        q
-    }
-
-    /// Convenience wrapper for single-MCU tests where host_time == tick as f64.
-    fn q_with(ring_depth: u32, starts: &[u64]) -> AxisQueue {
-        let pairs: Vec<(u64, f64)> = starts.iter().map(|&s| (s, s as f64)).collect();
-        q_with_host(ring_depth, &pairs)
-    }
-
-    #[test]
-    fn idle_when_empty() {
-        let queues: BTreeMap<AxisKey, AxisQueue> = BTreeMap::new();
-        assert!(matches!(schedule(&queues, 255, |_| None), Schedule::Idle));
-    }
-
-    #[test]
-    fn stalls_when_global_head_ring_full() {
-        let mut queues = BTreeMap::new();
-        let mut a = q_with(2, &[10]);
-        a.pushed = 2; // full
-        queues.insert(AxisKey { mcu_id: 1, axis: 0 }, a);
-        queues.insert(AxisKey { mcu_id: 2, axis: 0 }, q_with(8, &[20]));
-        assert!(matches!(
-            schedule(&queues, 255, |_| None),
-            Schedule::StallFull(AxisKey { mcu_id: 1, axis: 0 })
-        ));
-    }
-
-    #[test]
-    fn batches_contiguous_same_mcu_prefix_only() {
-        let mut queues = BTreeMap::new();
-        queues.insert(AxisKey { mcu_id: 1, axis: 0 }, q_with(8, &[0, 3]));
-        queues.insert(AxisKey { mcu_id: 1, axis: 1 }, q_with(8, &[1]));
-        queues.insert(AxisKey { mcu_id: 2, axis: 0 }, q_with(8, &[2]));
-        let s = schedule(&queues, 255, |_| None);
-        match s {
-            Schedule::Send(frames) => {
-                let ax: Vec<_> = frames.iter().map(|f| (f.key, f.pieces.len())).collect();
-                assert!(ax.contains(&(AxisKey { mcu_id: 1, axis: 0 }, 1)));
-                assert!(ax.contains(&(AxisKey { mcu_id: 1, axis: 1 }, 1)));
-                assert!(!ax.iter().any(|(k, _)| k.mcu_id == 2));
-            }
-            other => panic!("expected Send, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn frame_cap_splits() {
-        let mut queues = BTreeMap::new();
-        queues.insert(AxisKey { mcu_id: 1, axis: 0 }, q_with(8, &[0, 1, 2, 3]));
-        let s = schedule(&queues, 2, |_| None);
-        match s {
-            Schedule::Send(frames) => {
-                assert_eq!(frames.len(), 1);
-                assert_eq!(frames[0].pieces.len(), 2);
-            }
-            other => panic!("expected Send, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn full_axis_does_not_block_same_mcu_sibling() {
-        let mut q: BTreeMap<AxisKey, AxisQueue> = BTreeMap::new();
-        let yq = q_with(8, &[0, 2]);
-        let mut xq = q_with(1, &[1]);
-        xq.pushed = 1;
-        q.insert(AxisKey { mcu_id: 1, axis: 1 }, yq);
-        q.insert(AxisKey { mcu_id: 1, axis: 0 }, xq);
-        match schedule(&q, 255, |_| None) {
-            Schedule::Send(frames) => {
-                let yf = frames
-                    .iter()
-                    .find(|f| f.key == AxisKey { mcu_id: 1, axis: 1 });
-                assert!(yf.is_some(), "Y should be batched despite full sibling X");
-                assert!(
-                    !frames
-                        .iter()
-                        .any(|f| f.key == AxisKey { mcu_id: 1, axis: 0 }),
-                    "full X must not appear in the batch"
-                );
-            }
-            other => panic!("expected Send, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn time_gate_blocks_piece_beyond_horizon() {
-        let mut queues = BTreeMap::new();
-        queues.insert(AxisKey { mcu_id: 1, axis: 0 }, q_with(8, &[100]));
-        queues.insert(AxisKey { mcu_id: 1, axis: 1 }, q_with(8, &[200]));
-        match schedule(&queues, 255, |_| Some(150)) {
-            Schedule::Send(frames) => {
-                assert_eq!(frames.len(), 1, "only axis 0 should be batched");
-                assert_eq!(frames[0].key, AxisKey { mcu_id: 1, axis: 0 });
-                assert_eq!(frames[0].pieces.len(), 1);
-            }
-            other => panic!("expected Send, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn all_beyond_horizon_returns_stall_ahead() {
-        let mut queues = BTreeMap::new();
-        queues.insert(AxisKey { mcu_id: 1, axis: 0 }, q_with(8, &[1000]));
-        assert!(
-            matches!(
-                schedule(&queues, 255, |_| Some(500)),
-                Schedule::StallAhead(AxisKey { mcu_id: 1, axis: 0 })
-            ),
-            "expected StallAhead when sole piece is beyond horizon"
-        );
-    }
-
-    #[test]
-    fn no_horizon_none_uses_count_only_gate() {
-        let mut queues = BTreeMap::new();
-        queues.insert(AxisKey { mcu_id: 1, axis: 0 }, q_with(8, &[u64::MAX]));
-        match schedule(&queues, 255, |_| None) {
-            Schedule::Send(frames) => {
-                assert_eq!(frames.len(), 1);
-                assert_eq!(frames[0].pieces.len(), 1);
-            }
-            other => panic!("expected Send (no time gate), got {other:?}"),
-        }
-    }
-
-    /// Regression: bench shape from 2026-06-04.
-    ///
-    /// mcu1 (F446, 180 MHz) has a far-future piece whose tick value is numerically
-    /// small (~4.8e12) but whose host time is large (beyond mcu1's horizon).
-    /// mcu0 (H7, 520 MHz) has past-due pieces whose tick value is numerically large
-    /// (~13.8e12) but whose host time is now-ish (within mcu0's horizon, room available).
-    ///
-    /// Old code: `min_by` on raw ticks → mcu1's small ticks win → StallAhead(mcu1),
-    /// H7 starved for up to 2+ seconds.
-    /// New code: `min_by` on host time → mcu0's smaller host time wins → Send(mcu0).
-    #[test]
-    fn cross_mcu_host_time_ordering_bench_regression() {
-        let f446_tick: u64 = 4_790_000_000_000;
-        let h7_tick: u64 = 13_800_000_000_000;
-
-        // mcu1 (F446): numerically smaller tick, but host time is far in the future
-        let f446_host: f64 = 1_000.0; // far future, beyond any horizon
-        // mcu0 (H7): numerically larger tick, but host time is now-ish
-        let h7_host: f64 = 1.0; // near present
-
-        let mut queues = BTreeMap::new();
-        queues.insert(
-            AxisKey { mcu_id: 1, axis: 2 }, // F446, Z axis
-            q_with_host(8, &[(f446_tick, f446_host)]),
-        );
-        queues.insert(
-            AxisKey { mcu_id: 0, axis: 0 }, // H7, X axis
-            q_with_host(8, &[(h7_tick, h7_host)]),
-        );
-
-        // mcu0 horizon covers h7_tick; mcu1 horizon does NOT cover f446_tick.
-        let horizon_of = |mcu_id: u32| -> Option<u64> {
-            if mcu_id == 0 {
-                Some(h7_tick + 1_000_000)
-            } else {
-                Some(f446_tick - 1) // mcu1 piece is ahead of this horizon
-            }
-        };
-
-        match schedule(&queues, 255, horizon_of) {
-            Schedule::Send(frames) => {
-                assert_eq!(frames.len(), 1);
-                assert_eq!(
-                    frames[0].key.mcu_id, 0,
-                    "H7 (mcu0) should be selected, not F446 (mcu1)"
-                );
-            }
-            other => panic!(
-                "expected Send(mcu0) — cross-MCU host-time ordering regression, got {other:?}"
-            ),
-        }
-    }
-
-    /// The globally host-earliest piece's queue having room==0 must block everything
-    /// (intentional global-gating invariant).
-    #[test]
-    fn stall_full_on_globally_earliest_gates_all() {
-        let mut queues = BTreeMap::new();
-
-        // mcu0: host-earliest piece, but ring is full
-        let mut mcu0_q = q_with_host(2, &[(100, 1.0)]);
-        mcu0_q.pushed = 2; // full
-        queues.insert(AxisKey { mcu_id: 0, axis: 0 }, mcu0_q);
-
-        // mcu1: host-later piece, ring has room
-        queues.insert(AxisKey { mcu_id: 1, axis: 0 }, q_with_host(8, &[(50, 5.0)]));
-
-        assert!(
-            matches!(
-                schedule(&queues, 255, |_| None),
-                Schedule::StallFull(AxisKey { mcu_id: 0, axis: 0 })
-            ),
-            "StallFull on the globally host-earliest queue must gate all issuance"
-        );
-    }
-}
+mod sched_tests;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::mpsc;
-    use std::sync::{Arc, Mutex};
-
-    #[test]
-    fn room_full_then_drains() {
-        let mut q = AxisQueue::new(4);
-        assert_eq!(q.room(), 4);
-        q.pushed = 4;
-        assert_eq!(q.room(), 0); // full
-        q.retired = 1;
-        assert_eq!(q.room(), 1); // one freed
-    }
-
-    #[test]
-    fn room_correct_across_u32_wrap() {
-        let mut q = AxisQueue::new(8);
-        q.pushed = 2;
-        q.retired = u32::MAX;
-        assert_eq!(q.room(), 5);
-    }
-
-    #[test]
-    fn physical_write_cursor_advances_and_wraps_at_n() {
-        let mut q = AxisQueue::new(4);
-        assert_eq!(q.physical_write_cursor, 0);
-        q.advance_write_cursor(3);
-        assert_eq!(q.physical_write_cursor, 3);
-        q.advance_write_cursor(3);
-        assert_eq!(q.physical_write_cursor, 2);
-    }
-
-    #[derive(Clone)]
-    struct RecordingSink {
-        calls: Arc<Mutex<Vec<(u16, u32)>>>,
-    }
-
-    impl RecordingSink {
-        fn new() -> Self {
-            Self {
-                calls: Arc::new(Mutex::new(Vec::new())),
-            }
-        }
-        fn recorded(&self) -> Vec<(u16, u32)> {
-            self.calls.lock().unwrap().clone()
-        }
-    }
-
-    impl PieceSink for RecordingSink {
-        fn send_frame(
-            &self,
-            _key: AxisKey,
-            _pieces: &[PieceEntry],
-            start_slot: u16,
-            new_head: u32,
-        ) -> Result<i32, SendError> {
-            self.calls.lock().unwrap().push((start_slot, new_head));
-            Ok(kalico_protocol::result_codes::OK)
-        }
-    }
-
-    fn make_piece(t: u64) -> (PieceEntry, f64) {
-        (
-            PieceEntry {
-                start_time: t,
-                coeffs: [0.0; 4],
-                duration: 0.001,
-                _reserved: 0,
-            },
-            t as f64,
-        )
-    }
-
-    #[test]
-    fn run_pump_sets_start_slot_from_cursor_and_advances_it() {
-        const RING_DEPTH: u32 = 8;
-        const N: u32 = 3;
-
-        let sink = RecordingSink::new();
-        let (tx, rx) = mpsc::channel::<PumpMsg>();
-        let sink_clone = sink.clone();
-        let handle = std::thread::spawn(move || {
-            run_pump(rx, sink_clone, |_key| RING_DEPTH, |_mcu| None, |_| {});
-        });
-
-        tx.send(PumpMsg::Enqueue(EnqueueMsg {
-            key: AxisKey { mcu_id: 1, axis: 0 },
-            pieces: (0..N).map(|i| make_piece(i as u64)).collect(),
-            fresh_stream: false,
-        }))
-        .unwrap();
-        {
-            let deadline = std::time::Instant::now() + std::time::Duration::from_secs(2);
-            while sink.recorded().is_empty() {
-                assert!(
-                    std::time::Instant::now() < deadline,
-                    "pump did not drain first batch within deadline"
-                );
-                std::thread::yield_now();
-            }
-        }
-
-        tx.send(PumpMsg::Enqueue(EnqueueMsg {
-            key: AxisKey { mcu_id: 1, axis: 0 },
-            pieces: (N..N * 2).map(|i| make_piece(i as u64)).collect(),
-            fresh_stream: false,
-        }))
-        .unwrap();
-        {
-            let deadline = std::time::Instant::now() + std::time::Duration::from_secs(2);
-            while sink.recorded().len() < 2 {
-                assert!(
-                    std::time::Instant::now() < deadline,
-                    "pump did not drain second batch within deadline"
-                );
-                std::thread::yield_now();
-            }
-        }
-
-        tx.send(PumpMsg::Shutdown).unwrap();
-        handle.join().unwrap();
-
-        let recorded = sink.recorded();
-        assert_eq!(
-            recorded.len(),
-            2,
-            "expected exactly 2 sends, got {}",
-            recorded.len()
-        );
-
-        let (s0, h0) = recorded[0];
-        let (s1, h1) = recorded[1];
-
-        assert_eq!(s0, 0, "first start_slot should be 0");
-        assert_eq!(h0, N, "first new_head should be N={N}");
-
-        let expected_s1 = (N % RING_DEPTH) as u16;
-        assert_eq!(s1, expected_s1, "second start_slot should be {expected_s1}");
-        assert_eq!(h1, N * 2, "second new_head should be {}", N * 2);
-    }
-
-    #[test]
-    fn junction_jumps_math() {
-        // Exact gap: first piece starts exactly where previous ended.
-        let (tick_us, host_us) = junction_jumps(2000, 2.0e-3, 1000, 1.0e-3, 1_000_000.0);
-        assert!((tick_us - 1000.0).abs() < 1e-6, "tick_jump_us={tick_us}");
-        assert!((host_us - 1000.0).abs() < 1e-6, "host_jump_us={host_us}");
-
-        // Overlap (negative jump).
-        let (tick_us2, host_us2) = junction_jumps(900, 0.9e-3, 1000, 1.0e-3, 1_000_000.0);
-        assert!(tick_us2 < 0.0, "overlap should be negative tick jump");
-        assert!(host_us2 < 0.0, "overlap should be negative host jump");
-
-        // Cross-domain divergence: tick gap == 0 µs but host gap == 500 µs.
-        let freq = 520_000_000.0_f64;
-        let prev_end_ticks: u64 = 10_000;
-        let (tick_us3, host_us3) =
-            junction_jumps(prev_end_ticks, 5.0e-4, prev_end_ticks, 0.0, freq);
-        assert!(
-            (tick_us3).abs() < 1e-6,
-            "tick gap should be zero, got {tick_us3}"
-        );
-        assert!((host_us3 - 500.0).abs() < 1e-3, "host_jump_us={host_us3}");
-    }
-}
+mod tests;
 
 pub struct EnqueueMsg {
     pub key: AxisKey,
@@ -1043,82 +660,4 @@ impl PieceSink for WireSink {
 }
 
 #[cfg(test)]
-mod wire_sink_tests {
-    use super::*;
-    use kalico_host_rt::transport::TransportError;
-    use std::os::unix::net::UnixStream;
-    use std::sync::Arc;
-    use std::time::Duration;
-
-    /// A `UnixNativeConn` whose peer end is dropped — every call returns
-    /// `TransportError::Closed` (read returns `Ok(0)`) or `TransportError::Io`
-    /// (write to closed peer yields BrokenPipe).
-    fn closed_conn() -> Arc<kalico_host_rt::unix_native_conn::UnixNativeConn> {
-        let (client, _server) = UnixStream::pair().unwrap();
-        // Drop _server immediately: the client will see EOF / broken-pipe on
-        // the very first I/O.
-        Arc::new(kalico_host_rt::unix_native_conn::UnixNativeConn::from_stream(client))
-    }
-
-    fn key() -> AxisKey {
-        AxisKey { mcu_id: 0, axis: 0 }
-    }
-
-    fn one_piece() -> Vec<runtime::piece_ring::PieceEntry> {
-        vec![runtime::piece_ring::PieceEntry {
-            start_time: 1000,
-            coeffs: [0.0; 4],
-            duration: 0.001,
-            _reserved: 0,
-        }]
-    }
-
-    #[test]
-    fn closed_peer_yields_fatal_send_error() {
-        let sink = WireSink {
-            transports: {
-                let mut m = HashMap::new();
-                m.insert(0, McuTransport::EtherCat(closed_conn()));
-                m
-            },
-            timeout: Duration::from_millis(50),
-        };
-        let pieces = one_piece();
-        match sink.call_push_pieces(key(), &pieces, 0, 1) {
-            Err(SendError::Fatal(_)) => {}
-            other => panic!("expected Fatal for closed EtherCAT peer, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn timeout_yields_transient_send_error() {
-        // Construct a TransportError::Timeout and verify the match arm
-        // classifies it as Transient.  We can't easily produce a real
-        // timeout without sleeping, so test the classification logic
-        // directly on the error type.
-        let e = TransportError::Timeout;
-        let is_fatal = matches!(e, TransportError::Closed | TransportError::Io(_));
-        assert!(!is_fatal, "Timeout must not be fatal");
-    }
-
-    #[test]
-    fn parse_error_yields_transient_send_error() {
-        let e = TransportError::Parse("bad frame".to_owned());
-        let is_fatal = matches!(e, TransportError::Closed | TransportError::Io(_));
-        assert!(!is_fatal, "Parse must not be fatal");
-    }
-
-    #[test]
-    fn io_error_yields_fatal_send_error() {
-        let e = TransportError::Io(std::io::Error::from(std::io::ErrorKind::BrokenPipe));
-        let is_fatal = matches!(e, TransportError::Closed | TransportError::Io(_));
-        assert!(is_fatal, "Io must be fatal");
-    }
-
-    #[test]
-    fn closed_variant_is_fatal() {
-        let e = TransportError::Closed;
-        let is_fatal = matches!(e, TransportError::Closed | TransportError::Io(_));
-        assert!(is_fatal, "Closed must be fatal");
-    }
-}
+mod wire_sink_tests;
