+++ rust/motion-bridge/src/pump/tests.rs
@@ -1,0 +1,163 @@
+use super::*;
+use std::sync::mpsc;
+use std::sync::{Arc, Mutex};
+
+#[test]
+fn room_full_then_drains() {
+    let mut q = AxisQueue::new(4);
+    assert_eq!(q.room(), 4);
+    q.pushed = 4;
+    assert_eq!(q.room(), 0); // full
+    q.retired = 1;
+    assert_eq!(q.room(), 1); // one freed
+}
+
+#[test]
+fn room_correct_across_u32_wrap() {
+    let mut q = AxisQueue::new(8);
+    q.pushed = 2;
+    q.retired = u32::MAX;
+    assert_eq!(q.room(), 5);
+}
+
+#[test]
+fn physical_write_cursor_advances_and_wraps_at_n() {
+    let mut q = AxisQueue::new(4);
+    assert_eq!(q.physical_write_cursor, 0);
+    q.advance_write_cursor(3);
+    assert_eq!(q.physical_write_cursor, 3);
+    q.advance_write_cursor(3);
+    assert_eq!(q.physical_write_cursor, 2);
+}
+
+#[derive(Clone)]
+struct RecordingSink {
+    calls: Arc<Mutex<Vec<(u16, u32)>>>,
+}
+
+impl RecordingSink {
+    fn new() -> Self {
+        Self {
+            calls: Arc::new(Mutex::new(Vec::new())),
+        }
+    }
+    fn recorded(&self) -> Vec<(u16, u32)> {
+        self.calls.lock().unwrap().clone()
+    }
+}
+
+impl PieceSink for RecordingSink {
+    fn send_frame(
+        &self,
+        _key: AxisKey,
+        _pieces: &[PieceEntry],
+        start_slot: u16,
+        new_head: u32,
+    ) -> Result<i32, SendError> {
+        self.calls.lock().unwrap().push((start_slot, new_head));
+        Ok(kalico_protocol::result_codes::OK)
+    }
+}
+
+fn make_piece(t: u64) -> (PieceEntry, f64) {
+    (
+        PieceEntry {
+            start_time: t,
+            coeffs: [0.0; 4],
+            duration: 0.001,
+            _reserved: 0,
+        },
+        t as f64,
+    )
+}
+
+#[test]
+fn run_pump_sets_start_slot_from_cursor_and_advances_it() {
+    const RING_DEPTH: u32 = 8;
+    const N: u32 = 3;
+
+    let sink = RecordingSink::new();
+    let (tx, rx) = mpsc::channel::<PumpMsg>();
+    let sink_clone = sink.clone();
+    let handle = std::thread::spawn(move || {
+        run_pump(rx, sink_clone, |_key| RING_DEPTH, |_mcu| None, |_| {});
+    });
+
+    tx.send(PumpMsg::Enqueue(EnqueueMsg {
+        key: AxisKey { mcu_id: 1, axis: 0 },
+        pieces: (0..N).map(|i| make_piece(i as u64)).collect(),
+        fresh_stream: false,
+    }))
+    .unwrap();
+    {
+        let deadline = std::time::Instant::now() + std::time::Duration::from_secs(2);
+        while sink.recorded().is_empty() {
+            assert!(
+                std::time::Instant::now() < deadline,
+                "pump did not drain first batch within deadline"
+            );
+            std::thread::yield_now();
+        }
+    }
+
+    tx.send(PumpMsg::Enqueue(EnqueueMsg {
+        key: AxisKey { mcu_id: 1, axis: 0 },
+        pieces: (N..N * 2).map(|i| make_piece(i as u64)).collect(),
+        fresh_stream: false,
+    }))
+    .unwrap();
+    {
+        let deadline = std::time::Instant::now() + std::time::Duration::from_secs(2);
+        while sink.recorded().len() < 2 {
+            assert!(
+                std::time::Instant::now() < deadline,
+                "pump did not drain second batch within deadline"
+            );
+            std::thread::yield_now();
+        }
+    }
+
+    tx.send(PumpMsg::Shutdown).unwrap();
+    handle.join().unwrap();
+
+    let recorded = sink.recorded();
+    assert_eq!(
+        recorded.len(),
+        2,
+        "expected exactly 2 sends, got {}",
+        recorded.len()
+    );
+
+    let (s0, h0) = recorded[0];
+    let (s1, h1) = recorded[1];
+
+    assert_eq!(s0, 0, "first start_slot should be 0");
+    assert_eq!(h0, N, "first new_head should be N={N}");
+
+    let expected_s1 = (N % RING_DEPTH) as u16;
+    assert_eq!(s1, expected_s1, "second start_slot should be {expected_s1}");
+    assert_eq!(h1, N * 2, "second new_head should be {}", N * 2);
+}
+
+#[test]
+fn junction_jumps_math() {
+    // Exact gap: first piece starts exactly where previous ended.
+    let (tick_us, host_us) = junction_jumps(2000, 2.0e-3, 1000, 1.0e-3, 1_000_000.0);
+    assert!((tick_us - 1000.0).abs() < 1e-6, "tick_jump_us={tick_us}");
+    assert!((host_us - 1000.0).abs() < 1e-6, "host_jump_us={host_us}");
+
+    // Overlap (negative jump).
+    let (tick_us2, host_us2) = junction_jumps(900, 0.9e-3, 1000, 1.0e-3, 1_000_000.0);
+    assert!(tick_us2 < 0.0, "overlap should be negative tick jump");
+    assert!(host_us2 < 0.0, "overlap should be negative host jump");
+
+    // Cross-domain divergence: tick gap == 0 µs but host gap == 500 µs.
+    let freq = 520_000_000.0_f64;
+    let prev_end_ticks: u64 = 10_000;
+    let (tick_us3, host_us3) = junction_jumps(prev_end_ticks, 5.0e-4, prev_end_ticks, 0.0, freq);
+    assert!(
+        (tick_us3).abs() < 1e-6,
+        "tick gap should be zero, got {tick_us3}"
+    );
+    assert!((host_us3 - 500.0).abs() < 1e-3, "host_jump_us={host_us3}");
+}
